Screenshot: Refactor ScreenshotManager.save_image
In previous versions of GNOME Settings daemon (elementary OS <= 5),
saving to clipboard was handled by the daemon. The daemon used to
request Gala to save the screenshot in a temp path like
"/tmp/gnome-settings-daemon-screenshot-XXXX" and copied the image to
the  clipboard from there.

More information: https://github.com/elementary/gala/issues/474

Now, Gala is in charge of copying to clipboard when the take screenshot
methods receive an empty filename.

This commit: Move the code from save_image to save_image_to_file.