screenshot: Play shutter sound
We currently rely on Screenshot to play the shutter sound. This means
that it needs pulseaudio permissions in Flatpak and depending on the app
taking the screenshot you might not get a sound when a screenshot is
taken.

In addition, no sound was played when a screenshot was copied directly
to the clipboard.

Play the shutter sound in Gala to avoid this problems.