Dialogs: remove dbus timeout from AccessDialog call
we don't want the calls to the portal to be timed out before the user sent a
response, so disable the timeout.

while here, fix a issue with the show() call in the CloseDialog class creating
new dialogs even when a dialog is already open, also handle the portal response
only in the AccessDialog base class and use a protected method to convert
between portal and close dialog responses.

fixes https://github.com/elementary/gala/issues/1505