WindowManager: confirm_display_change() fixes
the value of 30 seconds here is wrong, and the dialog was outliving the timeout
used by mutter internally. instead of hardcoding the right values here too, use
the mutter provided function to query the real timeout.

also, make AccessDialog.show() base handler run first than signal callbacks,
and that we only call AccesDialog.close() when we don't receive a response.