Gestures: Add more details and introduce on_gesture_handled
Add information about whether a gesture will be handled. This might be used to
determine whether events are supposed to be consumed.
Add information about the origin of a gesture. This might be used to determine
whether a gesture should be handled (e.g. only allow swipe from screen edge).
Add common GestureBackend interface that all backends have to implement.
Add a on_gesture_handled signal to the GestureTracker which runs if and only if
true was returned in on_gesture_detected. Before it is emitted prepare_gesture_handling
will be called on the backend. This will be needed for the upcoming PanBackend to avoid
conflicts in event consuming.
Add a utility get_action to the GestureSettings that gets the corresponding action for a gesture.
This allows to remove some duplicated code.