ShellWindow: Disable unredirection when visible
Shell windows are in the shell group and are therefore above
fullscreen windows. They can also be visible when a window
is fullscreen via the new reveal feature. Also the animation plays
even when a window is already full screening. So we want to make
sure to disable unredirection when we are visible so that we are
actually visible.