WindowClone: Use a clone as drag handle (#2537)
Prerequisite for using a listmodel for the windows.
Also will allow us smoother transitions between workspaces in the future.
Saves us some complexity because we don't have to deal with our window clone being unparented.

This keeps our spot reserved when starting a drag and doesn't reflow all windows which IMO is better since there is now less happening when starting a drag which is easier to understand and on the eyes IMO. Also it is more inline with how other drag and drop works (e.g. dock, files, etc)
However if that's really not wanted it's easy to add that back.