ShellWindow: Only handle common code
Move things that only apply to PanelWindows there. Also rely on
subclasses to provide a GestureTarget for property updates instead
of trying to handle all cases.
This cleans up a lot of code that was necessary to handle the stuff
needed by all subclass.
This makes sure we only update things when actually needed. For example
ExtendedBehaviorWindows never need to update their target.
This will also provide more flexibility in the future if we want to
introduce other specialized windows. This also allows to change more
stuff in subclasses without affecting other subclasses.