DragDropAction: Don't consume events before drag started
We don't want to consume events let alone grab the actor
before we actually are sure that we will handle this drag.
This will improve situations where multiple gestures
can be started from the same position most notably
in the multitasking view drag and drop of window
clones and (once the new touchbackend is merged) swipe
left and right for switching workspaces.
This will also allow to use a separate click action instead
of having to implement clicking in the dnd action as well.