HideTracker: Assume hovered when revealing via barrier
In the case that the user triggered the barrier at a point
where the panel will actually reveal this is true anyways.
And in the case that the user triggered the barrier at a point
where the panel won't reveal because it's not wide enough
it will make sure that we recheck and unreveal it
when the pointer is invalidated.