Gala.Text: Don't extend Clutter.Text (#2680)
This seems to mess with... something. At least the picking (or at least something with the event delivery) was wrong.
That was noticeable in the multitasking view where the window clones have tooltips. Sometimes these tooltips (that use gala text) caused the window clones to not receive events most notably enter and leave as well as button press events.

I think this is the same issues that caused the hover widgets to sometimes not appear (see #2641) at least this PR now also fixes the enter/leave events.

Fixes https://github.com/elementary/gala/issues/2676